Add listener system to CompositeStorage.
If we want to support composites in composites (the case of storage channel (composite) -containing- disk drive (composite) -containing- disks), the parent composite needs to know when the child composite has a new source.

In practical terms, the disk drive composite needs to be able to let the storage channel composite know when a disk has been changed.