Use StorageProvider system in StorageNetworkNode as well.
Before, we tried to avoid this because it broke down in case of activeness updates (would cause duplicate storages in the storage channel, one initial storage and a duplicate when activeness is on).

However, we need this to properly handle network merges/updates or network node removals.

So now we expose a composite storage that can properly propagate changes to the root composite storage.