Interface I/O changes.
Ensure that an Interface that is acting as an External Storage can't extract from other Interfaces, to avoid I/O loops.
Lone interfaces can extract from anywhere, as they have no side-effects.
Just like in RS 1.
Self-extraction could be technically allowed but is disabled for now to be consistent with self-inserting behavior.

Ensure that an Interface that is acting as an External Storage can't insert into other Interfaces (or itself), to avoid insertion loops when clearing unwanted resources.
Self-inserting isn't allowed to avoid double counting momentarily during the insertion loop.