ci: no longer store version info in build.gradle
Per Semver, the version number being released depends on the changes in that release.

We can't predict those
changes at the start of a release cycle,
 so we can't bump the version at the start of a release cycle.

That means that the version number being released is determined at release time.

Because the version number is determined at release time,
we can't store any versioning metadata in the code.

If we did, build.gradle would have the version number of the latest released
version during the release cycle of the new version, which isn't correct.