fix: close grid when network changes
This is the easiest fix for now. In 8651f69, we attach to the storage channels of the new network.
However, this wouldn't sync the storages of the new network: we would have lingering grid stacks
that don't exist in the new network anymore.
Also, storage contents of the new network would only be visible
if the storages were added *after* the grid.

A better fix would be:

- Attach to the new storage channels.
- Add a priority system so the grid is added first - then storages.
This so that the grid can get notified in time about the added storages.
- On a network change, update the strategies for the new network.
- On a network change, clear the grid so lingering stacks will be gone.