fix: update network node activeness independently from the block state activeness
Previously, we would keep a cached "last active" value and update based on that. When it changed, we
would update the node activeness and propagate the block state change.

However, this setup breaks when the network node disconnects while the chunk is unloading.
How so? When the chunk unloads, it would update this "last active" value to "false", update the
network node activeness, and update the block state.
However, the block state updating would fail as
the chunk is currently unloading.
That means that when the network node disconnected,
activeness goes to "false", but the block state would stay "true".

Until now, this wasn't really an issue as we can safely
assume that the network node would become active again
as the chunk loads.

However, now we are adding wireless networking with the
network transmitter and network receiver. If a chunk
with a network receiver unloads, the network connection
would disconnect. If we rejoin the chunk, the activeness
on the block state would be "true", but there is no
network connection! It would only move the block state
to activeness "false" when a change occurs in the activeness,
as we only update the block state activeness when the network node activeness changes.

This change makes it so that we (still) update
the network node activeness, but, now are also allowed to update the block state later if we rejoin
a chunk to make it consistent.

This can cause a flicker at startup as the network boots up.
At startup, it detects that the block state is still active, however, the network
 is not initialized yet. It would send a block state
 update to become inactive and then another block state
 update to become active again as the network is initialized.

This wasn't an issue previously as the cached "last active"
was "false" at startup. Since activeness was also "false" at startup,
it wouldn't detect a change.
If the node goes to active a bit later, it would update the activeness in the node, but
a block state update wouldn't be necessary as it was active already. This essentially eliminated the
"initialization" block state update.

Overall, this change makes sure that activeness
is always consistent with the block state representation, even at startup, as a consequence.

This fixes the network receiver bug mentioned above, but
will also solve general possible block state inconsistency.