fix: network updating twice when activeness changes
This bug made it so that every time a network device is placed, it would trigger an extra update
after the initial update to connect
to an network.

This extra update was caused by the fact that the device was
going from inactive to active after the connection was made.

We now only update the network if the direction has changed, since that is the
only component that can influence graph scanning for now.

This bug also caused some flickering with the network transmitter
in combination with redstone mode.