fix: possible CME in GridView
From the linked issue a CME occurred at GridViewImpl#sort. It appeared like the backing list was
being modified while a sort was happening.

I noticed that the GridUpdatePacket runs on a different thread (had execute missing), so this change
makes it so that GridUpdatePacket handling should happen on the same thread as the sort, and should
avoid the CME happening in the first place.

I wasn't able to test this due to the nature of the bug.

See #480