fix: exporter crash if inserted is less than zero
This could happen if the remainder of the item
storage returns a higher amount
than originally inserted.
This was reproduceable with Functional Storage
1.4.1. Fill up a drawer, add an upgrade
to increase capacity, and remove the upgrade.
Next time you insert something, the drawer
will report a higher amount than allowed.
This hardens the exporter code a bit.

Also modifies CompositeStorageImpl to more
early crash when this happens.

Fixes #989