fix: storage disk info not updating properly
In a disk drive, it wouldn't update the
info of the second disk.
This is because the disk drive
constantly is asking for info
to render the progress bar.
However, when asking info about
the first disk, it would immediately
trigger the rate limiter.

Fix is to make a rate limiter per
storage ID.

Fixes #1109