fix: autocrafter deserialize failing if a pattern is invalid
If a crafting pattern becomes invalid
due to removing a mod,
the width and height would
mismatch the actual stack array list.
This would cause an ArrayIndexOutOfBoundsException
and the entire deserialize of the
autocrafter would fail.

If the pattern was in a player
inventory, the player couldn't log
in anymore.